WEEK2/Advance_RAG_bot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic
pydantic-settings
python-multipart==0.0.6
google-generativeai==0.3.0
groq==0.9.0
cohere==4.56.0
requests==2.31.0
pillow
sentence-transformers
faiss-cpu
pypdf==3.17.0
python-docx==1.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
transformers==4.35.2
torch
python-dotenv==1.0.0
pytesseract==0.3.10
PyPDF2
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22







